requirements.txt
asgiref==3.10.0
Django==4.2.25
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.15.0
